backend/requirements.txt
Django>=5.1.4
djangorestframework
django-cors-headers
python-dotenv
openai>=1.0.0
Pillow>=10.0.0
